make condition chain accept null as neutral boolean
